Renamed linear_i.h to .hpp.
