+++ src/solver/linear/linear_i.hpp
